Refactor UserRequestID to UserRequest, rename user to ApplicationUser, and adjust storage
