+++ ExpenseTracker/ExpenseTracker.Application/Mappings/UserMappings.cs
@@ -1,14 +1,14 @@
-using ExpenseTracker.Application.Requests.User;
-using ExpenseTracker.Application.ViewModels.User;
+using ExpenseTracker.Application.Requests.ApplicationUser;
+using ExpenseTracker.Application.ViewModels.ApplicationUser;
 using ExpenseTracker.Domain.Entities;
 
 namespace ExpenseTracker.Application.Mappings;
 
 public static class UserMappings
 {
-    public static UserViewModel ToViewModel(this ApplicationUser user)
+    public static ApplicationUserViewModel ToViewModel(this ApplicationUser user)
     {
-        return new UserViewModel
+        return new ApplicationUserViewModel()
         {
             Id = user.Id,
             UserName = user.UserName!,
@@ -17,17 +17,7 @@ public static class UserMappings
         };
     }
 
-    public static ApplicationUser ToEntity(this CreateUserRequest request)
-    {
-        return new ApplicationUser
-        {
-            UserName = request.UserName,
-            Email = request.Email,
-            PhoneNumber = request.PhoneNumber
-        };
-    }
-
-    public static ApplicationUser ToEntity(this UpdateUserRequest request)
+    public static ApplicationUser ToEntity(this UpdateApplicationUserRequest request)
     {
         return new ApplicationUser
         {
