+++ ExpenseTracker/ExpenseTracker.Application/Requests/Category/CategoryRequest.cs
@@ -5,4 +5,4 @@ namespace ExpenseTracker.Application.Requests.Category;
 public sealed record CategoryRequest(
     Guid UserId,
     int Id)
-    : UserRequestId(UserId: UserId);
+    : UserRequest(UserId: UserId);
