+++ ExpenseTracker/ExpenseTracker.Application/Requests/Category/CreateCategoryRequest.cs
@@ -8,4 +8,4 @@ public record CreateCategoryRequest(
     string Name,
     string? Description,
     CategoryType Type)
-    : UserRequestId(UserId: UserId);
+    : UserRequest(UserId: UserId);
