+++ ExpenseTracker/ExpenseTracker.Application/Requests/Category/GetCategoriesRequest.cs
@@ -3,4 +3,4 @@ using ExpenseTracker.Application.Requests.Common;
 namespace ExpenseTracker.Application.Requests.Category;
 
 public sealed record GetCategoriesRequest(Guid UserId,string? Search)
-    : UserRequestId(UserId: UserId);
+    : UserRequest(UserId: UserId);
