+++ ExpenseTracker/ExpenseTracker.Application/Requests/Common/UserRequest.cs
@@ -1,3 +1,3 @@
 ﻿namespace ExpenseTracker.Application.Requests.Common;
 
-public record UserRequestId(Guid UserId);
+public record UserRequest(Guid UserId);
