+++ ExpenseTracker/ExpenseTracker.Application/Requests/Notification/GetNotificationsRequest.cs
@@ -3,4 +3,4 @@
 namespace ExpenseTracker.Application.Requests.Notification;
 
 public sealed record GetNotificationsRequest(Guid UserId, int Something)
-    : UserRequestId(UserId: UserId);
+    : UserRequest(UserId: UserId);
