+++ ExpenseTracker/ExpenseTracker.Application/Requests/Transfer/TransferRequest.cs
@@ -3,4 +3,4 @@ using ExpenseTracker.Application.Requests.Common;
 namespace ExpenseTracker.Application.Requests.Transfer;
 
 public sealed record TransferRequest(Guid UserId, int Id) 
-    : UserRequestId(UserId: UserId);
+    : UserRequest(UserId: UserId);
