--- ExpenseTracker/ExpenseTracker.Application/Requests/User/CreateUserRequest.cs
@@ -1,9 +1,0 @@
-namespace ExpenseTracker.Application.Requests.User;
-
-public record CreateUserRequest(
-    Guid Id,
-    string UserName,
-    string Email,
-    string? PhoneNumber,
-    string Password)
-    : Common.UserRequestId(Id);
