--- ExpenseTracker/ExpenseTracker.Application/Requests/User/GetUserRequest.cs
@@ -1,9 +1,0 @@
-namespace ExpenseTracker.Application.Requests.User;
-
-public sealed record GetUserRequest(
-    Guid Id,
-    string UserName,
-    string Email,
-    string? PhoneNumber,
-    string? Search)
-    : Common.UserRequestId(Id);
