--- ExpenseTracker/ExpenseTracker.Application/Requests/User/UpdateUserRequest.cs
@@ -1,14 +1,0 @@
-namespace ExpenseTracker.Application.Requests.User;
-
-public sealed record UpdateUserRequest(
-    Guid Id,
-    string UserName,
-    string Email,
-    string? PhoneNumber,
-    string Password)
-    : CreateUserRequest(
-        Id,
-        UserName,
-        Email,
-        PhoneNumber,
-        Password);
