--- ExpenseTracker/ExpenseTracker.Application/Requests/User/UserRequest.cs
@@ -1,8 +1,0 @@
-namespace ExpenseTracker.Application.Requests.User;
-
-public sealed record UserRequest(
-    Guid Id,
-    string UserName,
-    string Email,
-    string? PhoneNumber)
-    : Common.UserRequestId(Id);
