+++ ExpenseTracker/ExpenseTracker.Application/Requests/Wallet/GetWalletsRequest.cs
@@ -3,4 +3,4 @@ using ExpenseTracker.Application.Requests.Common;
 namespace ExpenseTracker.Application.Requests.Wallet;
 
 public sealed record GetWalletsRequest(Guid UserId, string? Search) 
-    : UserRequestId(UserId: UserId);
+    : UserRequest(UserId: UserId);
