+++ ExpenseTracker/ExpenseTracker.Application/Requests/Wallet/WalletRequest.cs
@@ -3,4 +3,4 @@ using ExpenseTracker.Application.Requests.Common;
 namespace ExpenseTracker.Application.Requests.Wallet;
 
 public sealed record WalletRequest(Guid UserId, int Id) 
-    : UserRequestId(UserId: UserId);
+    : UserRequest(UserId: UserId);
