+++ ExpenseTracker/ExpenseTracker.Application/Requests/WalletShare/GetWalletSharesRequest.cs
@@ -3,4 +3,4 @@
 namespace ExpenseTracker.Application.Requests.WalletShare;
 
 public sealed record GetWalletSharesRequest(Guid UserId, string? Search)
-    : UserRequestId(UserId: UserId);
+    : UserRequest(UserId: UserId);
