+++ ExpenseTracker/ExpenseTracker.Application/Requests/WalletShare/WalletShareRequest.cs
@@ -3,4 +3,4 @@
 namespace ExpenseTracker.Application.Requests.WalletShare;
 
 public sealed record WalletShareRequest(Guid UserId, int Id)
-    : UserRequestId(UserId: UserId);
+    : UserRequest(UserId: UserId);
