+++ ExpenseTracker/ExpenseTracker.Application/Stores/Interfaces/INotificationStore.cs
@@ -1,6 +1,5 @@
 ﻿using ExpenseTracker.Application.Requests.Common;
 using ExpenseTracker.Application.Requests.Notification;
-using ExpenseTracker.Application.Requests.User;
 using ExpenseTracker.Application.ViewModels.Notification;
 
 namespace ExpenseTracker.Application.Stores.Interfaces;
@@ -8,5 +7,5 @@ namespace ExpenseTracker.Application.Stores.Interfaces;
 public interface INotificationStore
 {
     List<NotificationViewModel> GetAll(GetNotificationsRequest request);
-    int CountNotifications(UserRequestId request);
+    int CountNotifications(UserRequest request);
 }
