+++ ExpenseTracker/ExpenseTracker.Application/Stores/Interfaces/IUserStore.cs
@@ -1,15 +1,8 @@
-using ExpenseTracker.Application.Requests.Common;
-using ExpenseTracker.Application.Requests.User;
-using ExpenseTracker.Application.ViewModels.User;
-using Microsoft.AspNetCore.Http;
+using ExpenseTracker.Application.Requests.ApplicationUser;
 
 namespace ExpenseTracker.Application.Stores.Interfaces;
 
 public interface IUserStore
 {
-    Task<List<UserViewModel>> GetAll(GetUserRequest request);
-    Task<UserViewModel> GetById(UserRequest request);
-    Task<UserViewModel> Create(CreateUserRequest request, IEnumerable<IFormFile> attachments);
-    Task Update(UpdateUserRequest request);
-    Task Delete(UserRequest request);
+    Task Update(UpdateApplicationUserRequest request);
 }
