+++ ExpenseTracker/ExpenseTracker/Controllers/TransfersController.cs
@@ -53,7 +53,7 @@ public class TransfersController : Controller
         return View(transfer);
     }
 
-    public IActionResult Create([FromHeader] UserRequestId request)
+    public IActionResult Create([FromHeader] UserRequest request)
     {
         PopulateViewBag(request);
 
@@ -183,7 +183,7 @@ public class TransfersController : Controller
         return true;
     }
 
-    private void PopulateViewBag(UserRequestId request, int? categoryId = null, int? walletId = null)
+    private void PopulateViewBag(UserRequest request, int? categoryId = null, int? walletId = null)
     {
         var categories = _categoryStore.GetAll(new GetCategoriesRequest(request.UserId, null));
         var wallets = _walletStore.GetAll(new GetWalletsRequest(request.UserId, null));
